+++ hal/include/HAL/cpp/Log.h
