+++ wpilibj/src/athena/cpp/lib/HAL.cpp
@@ -9,7 +9,7 @@
 
 #include <assert.h>
 #include <jni.h>
-#include "Log.h"
+#include "HAL/cpp/Log.h"
 
 #include "HAL/HAL.h"
 #include "edu_wpi_first_wpilibj_hal_HAL.h"
