+++ wpilibj/src/athena/cpp/lib/I2CJNI.cpp
@@ -7,7 +7,7 @@
 
 #include <assert.h>
 #include <jni.h>
-#include "Log.h"
+#include "HAL/cpp/Log.h"
 
 #include "edu_wpi_first_wpilibj_hal_I2CJNI.h"
