+++ wpilibj/src/athena/cpp/lib/SolenoidJNI.cpp
@@ -8,7 +8,7 @@
 #include <jni.h>
 #include "HAL/HAL.h"
 #include "HAL/handles/HandlesInternal.h"
-#include "Log.h"
+#include "HAL/cpp/Log.h"
 
 #include "edu_wpi_first_wpilibj_hal_SolenoidJNI.h"
